# DINT-170 perf: remove metadata from project/repo list during onboard