Changes

github
Remove Active flag
Remove LastCommit sub-object

jira
Remove LastIssue sub-object
Remove TotalIssues field

Performance (pinpoint org)
github
onboard time 3.6s -> 2.3s
    + agent bin overhead 2.5s
jira
onboard time 4.8s -> 0.5s
    + agent bin overhead 2.5s